requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
fusepy==3.0.1
idna==3.4
python-dateutil==2.8.2
requests==2.28.1
six==1.16.0
tidalapi==0.7.0
urllib3==1.26.13
